requirements.txt
kaggle
transformers
torch 
torchvision 
torchaudio
pytorch_lightning
pillow